src/cfs/generate_tts_files/requirements.txt
google-api-python-client==1.7.9
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
google-cloud-texttospeech
google-cloud-storage
google-cloud-pubsub
functions-framework